refactor: change to hook
